refactor(api): Extract websocket_types.rs (ProgressEvent, ProgressBroadcaster)
- Create websocket_types.rs with ProgressEvent enum (10 variants) and ProgressBroadcaster
- Add 12 new unit tests for event serialization and broadcaster behavior
- websocket.rs: 427 -> 203 lines (-52% reduction)
- Total API lib tests: 339 (+9 from previous)
